backend/requirements.txt
numpy
scipy
websocket-client